Build script cleanup

+++ .world/clean.sh
@@ -2,4 +2,8 @@
 
 . .world/build_config.sh
 
+if [[ "$Linkage" == 'static' || ( "$Target" != 'linux' && "$Target" != 'windows_package' ) ]]; then
+  exit
+fi
+
 rm -rf python/out
